separate aggregated verify for orion simd field into a standalone file
