+++ poly_commit/src/orion.rs
@@ -26,6 +26,11 @@ pub use simd_field_impl::{
 #[cfg(test)]
 mod simd_field_tests;
 
+mod simd_field_agg_impl;
+
+#[cfg(test)]
+mod simd_field_agg_tests;
+
 mod pcs_for_expander_gkr;
 mod pcs_trait_impl;
 pub use pcs_trait_impl::{OrionBaseFieldPCS, OrionSIMDFieldPCS};
