+++ poly_commit/src/orion/simd_field_agg_tests.rs
@@ -1,0 +1,2 @@
+// TODO ... separate a large poly into parts, then separately commit/open,
+// finally aggregate and verify
